Fri, Sep 27, 2024, 11:17 AM -04:00

+++ runoff/runoff.c
@@ -173,7 +173,7 @@ bool print_winner(void)
     for (int i = 0; i < candidate_count; i++) {
         if (candidates[i].votes >= 1) {
             printf("Winning count is %i?\n", candidates[i].votes);
-            int x = voter_count/2;
+            float x = voter_count/2;
             printf("Total votes %f\n", x);
             return true;
         }
